Another round of bug/security fixes

+++ diesel_cli/tests/print_schema/print_schema_regression_test_for_2623/sqlite/schema.sql
@@ -6,7 +6,7 @@ CREATE TABLE tab_key1(
 
 CREATE TABLE tab_problem (
   id integer NOT NULL,
-  key1 bigint NOT NULL,
+  key1 integer NOT NULL,
 
   PRIMARY KEY (id),
   UNIQUE (key1),
